fix(global-header): fix crash in SoftwareTemplatesSection
