+++ workspaces/global-header/plugins/global-header/src/components/HeaderDropdownComponent/SoftwareTemplatesSection.tsx
@@ -14,12 +14,19 @@
  * limitations under the License.
  */
 
-import React from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 
+import { useApi } from '@backstage/core-plugin-api';
+import { catalogApiRef } from '@backstage/plugin-catalog-react';
+import { Entity } from '@backstage/catalog-model';
+
+import Box from '@mui/material/Box';
 import Divider from '@mui/material/Divider';
 import Typography from '@mui/material/Typography';
 
-import { MenuSection, MenuItemConfig } from './MenuSection';
+import { MenuSection } from './MenuSection';
+import { MenuItemLink } from '../MenuItemLink/MenuItemLink';
+import { ComponentType } from '../../types';
 
 /**
  * Software Templates Section properties
@@ -27,16 +34,60 @@ import { MenuSection, MenuItemConfig } from './MenuSection';
  * @public
  */
 export type SoftwareTemplatesSectionProps = {
-  items: MenuItemConfig[];
   handleClose: () => void;
   hideDivider?: boolean;
 };
 
 export const SoftwareTemplatesSection = ({
-  items,
   handleClose,
   hideDivider,
 }: SoftwareTemplatesSectionProps) => {
+  const catalogApi = useApi(catalogApiRef);
+
+  const [entities, setEntities] = useState<Entity[]>([]);
+  // TODO: handle loading
+  const [_loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+
+  useEffect(() => {
+    const fetchEntities = async () => {
+      try {
+        const response = await catalogApi.getEntities({
+          filter: { kind: ['Template'] },
+          limit: 7,
+        });
+        setEntities(response.items);
+      } catch (err) {
+        setError(err);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchEntities();
+  }, [catalogApi]);
+
+  const items = useMemo(() => {
+    return entities
+      .filter(e => e.kind === 'Template')
+      .map(m => ({
+        Component: MenuItemLink as React.ComponentType,
+        type: ComponentType.LINK,
+        label: m.metadata.title ?? m.metadata.name,
+        link: `/create/templates/default/${m.metadata.name}`,
+      }));
+  }, [entities]);
+
+  if (error) {
+    return (
+      <Box display="flex" justifyContent="center" alignItems="center" p={2}>
+        <Typography variant="body1" color="error">
+          Error fetching templates
+        </Typography>
+      </Box>
+    );
+  }
+
   if (items.length === 0) {
     return (
       <>
